server/requirements.txt
autopep8==1.3.4
colorama==0.3.9
decorator==4.2.1
Django==2.0.2
djangorestframework==3.7.7
Faker==0.8.11
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
parso==0.1.1
pep8==1.7.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
pycodestyle==2.3.1
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2018.3
rope==0.10.7
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
